# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Games API",
    "version": "1.0.0",
    "description": "Generated from src/server/api-contract.ts Valibot-backed endpoint contracts."
  },
  "paths": {
...
    "/api/multiplayer/socket": {
      "get": {
        "operationId": "connectMultiplayerSocket",
        "summary": "Upgrade to the multiplayer WebSocket protocol.",
        "tags": ["Multiplayer"],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "description": "Room code.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "playerId",
            "in": "query",
            "required": true,
            "description": "Session player id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "description": "Session bearer token.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "WebSocket upgrade succeeds."
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "ok": {
            "const": false
          },
          "error": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
    }
  }
}